backend/requirements-dev.txt
-r requirements.txt
pytest==8.*
pytest-flask==1.3.*
pytest-cov==6.*
factory-boy==3.3.*
ruff==0.9.*
